All endpoints work, but not all server pass-off tests work

+++ src/main/java/Services/FillService.java
@@ -1,6 +1,7 @@
 package Services;
 
 import DAO.*;
+import JSONReader.LocationData;
 import Model.*;
 import Result.FillResult;
 import java.sql.*;
@@ -13,6 +14,7 @@ public class FillService {
 
     private Connection connection;
     private Database db;
+    private LocationData locations;
 
     /**
      * Initializes an empty constructor for the class
@@ -21,17 +23,17 @@ public class FillService {
         db = new Database();
     }
 
-    /**
-     * Initializing constructor for class with a connection argument
-     *
-     * @param connection Links connection with database
-     */
-    public FillService(Connection connection) {
-        this.connection = connection;
-    }
+//    /**
+//     * Initializing constructor for class with a connection argument
+//     *
+//     * @param connection Links connection with database
+//     */
+//    public FillService(Connection connection) {
+//        this.connection = connection;
+//    }
 
     /**
-     * Fill methos to fill database with username and generations
+     * Fill methods to fill database with username and generations
      *
      * @param username
      * @param generations
@@ -41,30 +43,52 @@ public class FillService {
 
         FillResult response = new FillResult();
 
-
         try {
+            db.openConnection(); // open connection
 
-            db.openConnection();
             UserDao uDao = new UserDao(db.getConnection());
             EventDao eDao = new EventDao(db.getConnection());
             PersonDao pDao = new PersonDao(db.getConnection());
 
-            if(uDao.userExists(username)) {
+            if(uDao.userExists(username) && generations >= 0) {
 
+                User user = uDao.findUser(username);
 
-                pDao.clearPersonUsername(username);
-                eDao.deleteAllEvents(username);
+                pDao.clearPersonUsername(username); // clear all persons for user
+                eDao.deleteAllEvents(username); // clear all events for user
 
-                response.setMessage("Successfully added persons and events to the database!");
+                if(generations == 99) { // default generation lives here
+                    pDao.generateTree(
+                            user,
+                            user.getPersonID(),
+                            4,
+                            eDao);
+                } else { // any other number of generations lives here
+                    pDao.generateTree(
+                            user,
+                            user.getPersonID(),
+                            generations,
+                            eDao);
+                }
+
+                int numOfPersons = pDao.getNumOfPersons();
+                int numOfEvents = eDao.getNumOfEvents();
+
+                // Response message
+                response.setMessage("Successfully added " + numOfPersons + " persons and " + numOfEvents + " events to the database!");
 
                 response.setSuccess(true);
                 db.closeConnection(true);
+
             } else {
+
                 response.setSuccess(false);
-                response.setMessage("Error invalid userName or generations parameter - Fill Service");
+                response.setMessage("Error invalid username or generations - Fill Service");
                 db.closeConnection(false);
+
             }
         } catch(DataAccessException e) {
+
             response.setSuccess(false);
             response.setMessage("Internal server error - Fill Service");
 
@@ -72,9 +96,14 @@ public class FillService {
                 db.closeConnection(false);
             } catch(DataAccessException f) {
                 f.printStackTrace();
+            } catch (Exception j) {
+                j.printStackTrace();
             }
         }
 
         return response;
     }
+
+    // Generate data to populate the database
+
 }
